sicora-be-python/userservice/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.1.1
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
python-multipart==0.0.18
httpx==0.28.1
redis==5.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.3.0
greenlet
aiosqlite==0.20.0
